[AR-36] Update the navigation for the Dashboard routes
