+++ app/src/main/java/com/loraxx/electrick/autosweep/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.loraxx.electrick.autosweep
 
 import android.os.Bundle
-import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
@@ -19,7 +18,6 @@ import androidx.navigation3.ui.rememberSceneSetupNavEntryDecorator
 import com.loraxx.electrick.autosweep.navigation.Dashboard
 import com.loraxx.electrick.autosweep.navigation.Login
 import com.loraxx.electrick.autosweep.navigation.QuickBalance
-import com.loraxx.electrick.autosweep.ui.dashboard.ActionBeltItem
 import com.loraxx.electrick.autosweep.ui.dashboard.DashboardContainerScreen
 import com.loraxx.electrick.autosweep.ui.dashboard.DashboardViewModel
 import com.loraxx.electrick.autosweep.ui.login.LoginScreen
@@ -72,54 +70,7 @@ class MainActivity : ComponentActivity() {
                         }
                         entry<Dashboard> {
                             val dashboardViewModel: DashboardViewModel = hiltViewModel()
-                            DashboardContainerScreen(
-                                viewModel = dashboardViewModel,
-                                onTopUpClick = {
-                                    //TODO navigate to Top up screen
-                                },
-                                onHistoryClick = {
-                                    //TODO navigate to Transaction History screen
-                                },
-                                onActionBeltItemClick = { actionBeltItem ->
-                                    when (actionBeltItem) {
-                                        ActionBeltItem.RFID -> {
-                                            //TODO navigate to RFID screen
-                                            Toast.makeText(
-                                                this@MainActivity,
-                                                "Navigate to RFID screen",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        }
-
-                                        ActionBeltItem.TRAFFIC -> {
-                                            //TODO navigate to Traffic screen
-                                            Toast.makeText(
-                                                this@MainActivity,
-                                                "Navigate to Traffic screen",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        }
-
-                                        ActionBeltItem.TOLL_RATE -> {
-                                            //TODO navigate to Toll Rate screen
-                                            Toast.makeText(
-                                                this@MainActivity,
-                                                "Navigate to Toll Rate screen",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        }
-
-                                        ActionBeltItem.HELP -> {
-                                            //TODO navigate to Help screen
-                                            Toast.makeText(
-                                                this@MainActivity,
-                                                "Navigate to Help screen",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        }
-                                    }
-                                },
-                            )
+                            DashboardContainerScreen(viewModel = dashboardViewModel)
                         }
                     },
                     transitionSpec = {
@@ -136,7 +87,6 @@ class MainActivity : ComponentActivity() {
                     },
                 )
             }
-
         }
     }
 }
